fix: fixed sub_58490_radix_3d_2 to correct measure distance

+++ remc2/utilities/Maths.cpp
@@ -735,7 +735,7 @@ uint16_t Maths::x_WORD_DE350[257] = { //speed table3//2af350 tang?
 
 unsigned int Maths::sub_58490_radix_3d_2(axis_3d* a1, axis_3d* a2)//239490
 {
-	int16_t radix = ((int16_t)(a2->x - a1->x)) * ((int16_t)(a2->x - a1->x)) + ((int16_t)(a2->y - a1->y)) * ((int16_t)(a2->y - a1->y));
+	uint32_t radix = ((int16_t)(a2->x - a1->x)) * ((int16_t)(a2->x - a1->x)) + ((int16_t)(a2->y - a1->y)) * ((int16_t)(a2->y - a1->y));
 	return Maths::sub_7277A_radix_3d(radix);
 }
